pyproject.toml
[project]
name = "sfkm-me"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.11.5",
    "pygithub>=2.6.1",
]
